fix: store kardex movements as float64 to record fractional consumption
The inventory_movements columns quantity, stock_before and stock_after
were integers, so recipe explosion rounded fractional insumo consumption
before writing the kardex. Consuming 0.3 kg of Arroz logged quantity=-1,
stock_before=3, stock_after=2 — the audit trail lied about real stock
(violates Constitución Art. VII: el inventario es exacto).

- models.InventoryMovement: widen Quantity, StockBefore, StockAfter from
  int to float64; GORM maps them to numeric. AutoMigrate widens the
  legacy integer columns via a safe additive ALTER COLUMN ... TYPE
  numeric cast (Art. X).
- services.LogInventoryMovement: convert the integer product-stock
  values to float64 when building the movement; product stock stays
  integer-valued, only the movement trail widens.
- recipe_service.consumeIngredient: drop math.Ceil/Floor rounding and
  store exact values (Quantity: -consumed, StockBefore/StockAfter).
  Remove now-unused math import; simplify Notes.

No behavior change for product kardex or direct sales — AC-06 regression
test stays green. Movement-count assertions in the recipe/sale/sync
tests are unaffected.

+++ internal/models/inventory_movement.go
@@ -21,20 +21,26 @@ const (
 )
 
 type InventoryMovement struct {
-	ID             string       `gorm:"type:uuid;primaryKey" json:"id"`
-	TenantID       string       `gorm:"type:uuid;not null;index" json:"tenant_id"`
-	BranchID       *string      `gorm:"type:uuid;index" json:"branch_id,omitempty"`
-	ProductID      string       `gorm:"type:uuid;not null;index:idx_inv_mov_product_created,priority:1" json:"product_id"`
-	ProductName    string       `gorm:"type:varchar(256)" json:"product_name"`
-	MovementType   MovementType `gorm:"type:varchar(32);not null;index" json:"movement_type"`
-	Quantity       int          `gorm:"not null" json:"quantity"`
-	StockBefore    int          `gorm:"not null" json:"stock_before"`
-	StockAfter     int          `gorm:"not null" json:"stock_after"`
-	ReferenceID    *string      `gorm:"type:uuid" json:"reference_id,omitempty"`
-	ReferenceType  string       `gorm:"type:varchar(32)" json:"reference_type,omitempty"`
-	UserID         *string      `gorm:"type:uuid" json:"user_id,omitempty"`
-	UserName       string       `gorm:"type:varchar(128)" json:"user_name,omitempty"`
-	Notes          string       `gorm:"type:text" json:"notes,omitempty"`
-	IdempotencyKey *string      `gorm:"type:varchar(128);uniqueIndex:idx_inv_mov_idempotency,where:idempotency_key IS NOT NULL" json:"idempotency_key,omitempty"`
-	CreatedAt      time.Time    `gorm:"index:idx_inv_mov_product_created,priority:2,sort:desc" json:"created_at"`
+	ID           string       `gorm:"type:uuid;primaryKey" json:"id"`
+	TenantID     string       `gorm:"type:uuid;not null;index" json:"tenant_id"`
+	BranchID     *string      `gorm:"type:uuid;index" json:"branch_id,omitempty"`
+	ProductID    string       `gorm:"type:uuid;not null;index:idx_inv_mov_product_created,priority:1" json:"product_id"`
+	ProductName  string       `gorm:"type:varchar(256)" json:"product_name"`
+	MovementType MovementType `gorm:"type:varchar(32);not null;index" json:"movement_type"`
+	// Quantity, StockBefore and StockAfter are float64 so the kardex can
+	// record fractional consumption exactly (e.g. 0.3 kg of an insumo by
+	// recipe explosion). GORM maps them to a Postgres numeric column;
+	// AutoMigrate widens the legacy integer columns via a safe additive
+	// ALTER COLUMN ... TYPE numeric cast (Constitución Art. X). Product
+	// stock itself stays integer-valued — only the movement trail widens.
+	Quantity       float64   `gorm:"not null" json:"quantity"`
+	StockBefore    float64   `gorm:"not null" json:"stock_before"`
+	StockAfter     float64   `gorm:"not null" json:"stock_after"`
+	ReferenceID    *string   `gorm:"type:uuid" json:"reference_id,omitempty"`
+	ReferenceType  string    `gorm:"type:varchar(32)" json:"reference_type,omitempty"`
+	UserID         *string   `gorm:"type:uuid" json:"user_id,omitempty"`
+	UserName       string    `gorm:"type:varchar(128)" json:"user_name,omitempty"`
+	Notes          string    `gorm:"type:text" json:"notes,omitempty"`
+	IdempotencyKey *string   `gorm:"type:varchar(128);uniqueIndex:idx_inv_mov_idempotency,where:idempotency_key IS NOT NULL" json:"idempotency_key,omitempty"`
+	CreatedAt      time.Time `gorm:"index:idx_inv_mov_product_created,priority:2,sort:desc" json:"created_at"`
 }
